Abstracted tool away to use subcommands
